Move Format to its own file / submodule
